pyproject.toml
[tool.poetry]
name = "market-place-sprint"
version = "0.1.0"
description = ""
authors = ["igorynos <igorynos@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
sqlalchemy = "^2.0.31"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
unicorn = "^2.0.1.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
